+++ db/migrations/20240123135615_create_table_transactions.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE transactions (
     id BIGINT NOT NULL auto_increment,
-    total_amount DECIMAL(10, 2) NOT NULL,
+    total_amount INT NOT NULL,
     created_at  BIGINT    NOT NULL,
     updated_at  BIGINT    NOT NULL,
     PRIMARY KEY (id)
